+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Fixed implementation for ACL inheritance for Attribute
 * Fixed no permission check in advanced search and simple search(#282)
 * Fixed a different number of entries displayed on the entity dashboard (#308)
+* Fixed an issue where advanced search narrow down was slow (#321)
 
 ### Refactored
 * Refactored the entry list page